test: 회원가입 유효성 검사 테스트코드 구현
